Add procedural world generation

+++ src/umath.c
@@ -1,6 +1,6 @@
-#include "math/umath.h"
+#include "umath.h"
 
-double clamp(double val, double min, double max)
+float clamp(float val, float min, float max)
 {
     if (val < min) return min;
     if (val > max) return max;
